requirements-gpu.txt
asn1crypto==0.24.0
backcall==0.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography>=3.3.2
cycler==0.10.0
decorator==4.4.0
idna==2.8
imageio==2.5.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.0
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
# mkl-fft==1.0.12
# mkl-random==1.0.2
# mkl-service==2.0.2
numpy==1.16.4
olefile==0.46
pandas==0.25.0
#parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow>=6.2.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments>=2.7.4
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
#pyzmq==18.0.2
requests==2.22.0
scikit-learn==0.21.2
#scipy==1.3.0
seaborn==0.9.0
six==1.12.0
statsmodels==0.10.1
#torch==1.2.0
torchfile==0.1.0
#torchvision==0.4.2
#torchvision==0.5.0
torchvision==0.4.1
tornado==6.0.3
tqdm==4.33.0
traitlets==4.3.2
urllib3>=1.25.8
visdom==0.1.8.8
wcwidth==0.1.7
websocket-client==0.56.0
